python/requirements.txt
Flask
cfenv
sap-xssec
sap_cf_logging
hdbcli
statistics